+++ pkg/api/cloud/db/response.go
@@ -26,7 +26,7 @@ type (
 		Database models.Database `json:"return"`
 		models.APIResponse
 	}
-	// ListResponse is the returned response from the List endpoint.
+	// ListResponse is the returned response from the ListDomains endpoint.
 	ListResponse struct {
 		Return struct {
 			Databases []models.Database `json:"data"`